chore(extractors): separate product and product-tile extractors
