+++ src/extractors/__init__.py
@@ -1,1 +1,6 @@
-from .extractors import ColesProductExtractor, ColesProductTileExtractor
+"""
+Classes for extracting data from HTML content of Coles website.
+"""
+
+from .product_detail import ColesProductExtractor
+from .product_tile import ColesProductTileExtractor
